SXTranslationConfirmer: Remove assumption that section suggestion exists
Until recently, we always populated the "currentSectionSuggestion" state
variable, and we relied on this inside "Confirm a translation" step, to
compute several variables. However, since we changed the way we get the
current section suggestion in I84c8c5e70c1779585e3260a35366b8375c770532,
we should also adjust our assumptions about the section suggestion
being always non-empty inside this step: current section suggestion is
non-empty only for section translations (not page/article translations).

This wrong assumption leads to issues in desktop unified dashboard, as
the source title that is passed to "setCXToken" method is empty, and
the user is not redirected properly to the desktop editor.

This patch fixes this issue, by relying on the URL parameters to get
the current language pair and source page title, instead of relying
on the current section suggestion.

It also fixes the useActionPanel composable to follow the proper
composable syntax, and get the "currentSectionSuggestion" from the
corresponding composable, instead of expecting it as an argument.

Change-Id: I2ca0e8e5bb7105cf0b439f88a322bf4695a0db3f