SXTranslationConfirmerActionPanel: Rename useSectionSelectorClickHandler composable
The name of this composable is confusing, as the "Section Selector
(Pick a section)" is a different step than the "Confirm a translation",
and these click handlers do not necessarily select a specific section
to translate (actually this only happens when the section title is
prefilled within this step, which is not even the most usual case).

This patch renames this composable to a more meaningful and not confusing name.

Change-Id: I8cf7ebe110c3c34b12955103dcc2f1d16a9f057f