CX: Add dialog to notify users about fast unreviewed translations
This patch adds a dialog that informs the users that they have
a recent fast/unreviewed translations, that they should review
before starting a new translation. This dialog is displayed inside
"Confirm a translation" step, once the users click on any of the
action buttons of this step. A request to the "cxcheckunreviewed"
Action API endpoint is sent, to determine if such unreviewed
translation exists.

Codex dialog is used for the implementation of this dialog, and
this is the reason that some details may differ from the design
inside the task description. For now the Codex defaults have
been left as is, for consistency and simplicity.

Note: Minor eslint fix also added inside translator.js.

Bug: T331023
Change-Id: I69d6c38249a24ad17626e93ee029b96cca085acd