SXSectionSelector: Fix navigation back to dashboard
Currently, the navigation from "Pick a section" step to the dashboard is
broken. This happens because some SFCs inside this step expect the
current section suggestion object will always be populated. However, at
this point, the current section suggestion is computed based on the URL
parameters which are cleared before the redirection to dashboard.

That means, that for some milliseconds, while the navigation has not
been completed, and the "Pick a section" step is still displayed, the
current section suggestion can be empty.

This patch removes the assumptions for non-empty section suggestion from
these SFCs. It also uses the "useCurrentSectionSuggestion" to get the
current section object, instead of a prop, to avoid warnings about
wrong prop type, as these components expect a prop of SectionSuggestion
model, but the prop is defined, once the URL params are cleared.

Change-Id: Ia87b431ddcfe2b0183c7d260401eac2f05e466b5