SX: Go to confirmation step when starting SX from published translation
Until now, when the user selects to start a section translation from
a published translation of theirs, they were navigated directly to the
"Pick a section" step. However, now that the way we handle section
suggestions has changed, this leads to errors.

This patch navigates the user to the "Confirm a translation" step, when
the user starts such a section translation (either by clicking on the
first missing section title, or the ellipsis icon inside a published
translation card). Apart from fixing the errors, this way is consistent
with our workflow, since we always want to redirect the user to the
confirmation step, before starting a new translation.

Change-Id: I2dc222e78af6d06cc577d35f9a4b306fc2b12347