SX: Improve error handling for "fetchSegmentTranslation" API method
Currently, when segment translation request fails with an HTTP error
the promise inside "fetchSegmentTranslation" API method is not rejected.
Instead, "response.json" is resolved properly and the method ultimately
returns a promise that resolves to an undefined value.

On the other hand, when a network error occurs (e.g. if the request URL
is blocked), the "fetchSegmentTranslation" method returns a rejected
promise as expected.

The difference between those 2 behaviors above, is that "fetch" native
method doesn't throw an error when the request fails with an HTTP error,
but it only sets the "ok" property of its response to false.

To provide a consistent handling of failed requests, this patch handles
the response from "fetch" method, based on the "ok" property and returns
a rejected promise for both HTTP and network errors.

Additionally, the patch fixes the "translateContent" method to return
an empty string, instead of falling back to the original content, so
that the "RetryMtCard" component can be properly displayed.

Bug: T359900
Change-Id: I18e60513a7e041af46a953c58f22a807e2d97d9f