CX: Add mw.cx.UserPermissionChecker
Currently, we only check if user can publish in the main namespace,
upon translation initialization. However, the "Publish" button can
still be enabled, when user makes changes to the translation. Given
that we do not check again if user can publish, before publishing,
users without permissions end up being able to publish to the main
namespace, even if this should be blocked.

This patch moves the "checkIfUserCanPublish" and the related methods
from the mw.cx.init.Translation.js class to the newly created
mw.cx.UserPermissionChecker class, so that this functionality can
also be reused, when user clicks on the "Publish" button.

Bug: T349959
Change-Id: Ib253384df42d202b5d14b6a9fee930828361437a