MinT MVP: Add page title and adjust special page name
This patch renames the "Special:MinT" special page to
"Special:AutomaticTranslation", while keeping the "Special:MinT"
name as alias.

Additionally, the patch adds the page title, so that it is properly
displayed in the browser tab. Since the same page title is also
displayed inside the page content (which is not desired), the patch
adds some CSS rules to hide this title from the page content.

Bug: T364096
Change-Id: I3becfed2bf48492b44a6f76b79d8b8d40f33013d