MinT MVP: Re-fetch lead section translation on target language change
Currently the lead section translation inside the confirmation step
is only fetched during initialization, and the translation doesn't
change when the target language is updated.

To fix this issue, this patch refactors the confirmation step and
adds moves the lead section translation functionality to a new
composable. The method that the composable exposes, is called
every time the target language is updated.

Bug: T364730
Change-Id: I084141c5107047d604a058551a54135d946b1930