CX: Fix translation container max width for large screens
Inside mw.cx.ui.TranslationView.less file there is this rule that
defines the translation view container width for larger screens:

```
@media only screen and
  ( min-width: ( @wide + 0.25 * @max-tools-column-width ) ) {
      width: calc( 100% ~'-' @max-tools-column-width );
}
```

where @wide and @max-tools-column-width are defined inside
mw.cx.variables.less file:
```
@wide: 1200px;
@tools-column-margin: 20px;
@max-tools-column-width: ( @wide / 3 ) + @tools-column-margin;
```

The above LESS rules resolve to the following CSS rule:
```
@media only screen and (min-width: 1305px) {
    .cx-translation-view .cx-translation-view-container {
        width: calc(100% -(1200px / 3) + 20px);
    }
}
```
The above rule is not the expected, as given that the
@max-tools-column-width variable is calculated to "420px",
one would expect the rule for the width to be:
```width: calc(100% - 420px);```.

The current rule actually resolves to:
```width: calc(100% - 380px);```, which leads the tools container to
overflow, and be displayed below the translation container.

This patch fixes this issue by adding parentheses, so that the LESS
rule is properly compiled. It's still unknown why the previous
implementation stopped working, just now, but the patch is expected
to be enough for this regression to get fixed.

Bug: T366374
Change-Id: I423de88b58baddc0cf8562edce3a11499e120845
(cherry picked from commit 000f4796443b670fc8bccccbee835354b8dfcffe)