MinT: Simplify pageSearchResult sourceTitle getter
Until now the "sourceTitle" getter inside the pageSearchResult class,
was implemented in a similar way as the "getTitleByLanguage" method
of the same class, but missed to appropriately handle the case where
the "sourceLanguage" of the object is equal to its "language" property.

This patch fixes this issue and simplifies this getter by using the
"getTitleByLanguage" method to get the title for the source language.

Bug: T364621
Change-Id: Ic991532f4c7a146b0e2782ac4a6cfb536940fe59