AX language searcher entrypoint: Use configuration for enabled languages
The previous implementation basically supported the following scenario:
Enabling the configuration parameter for the AX entrypoint (i.e. the
"AutomaticTranslationEnableEntrypoints" config parameter) for Igbo
Wikipedia means that it will display the entrypoint card when searching
for another language on Igbo Wikipedia.

However, the above implementation is expected to have much less impact
than if we supported exposing the entrypoint in the language selector
of other languages when they search for Igbo.

This patch renames the "AutomaticTranslationEnableEntrypoints" config
to "AutomaticTranslationLanguageSearcherEntrypointEnabledLanguages",
and changes the implementation of the entrypoint to follow the new
logic.

Additionally, in this patch we also remove the usage of the display
language inside the "getMintUrl" method of the SiteMapper, since it
has been also removed from the AX URL params.

Bug: T363183
Change-Id: Ib7fd8a7bbf315101540e07b1fea6a906982aaa6c