CX: Fix "wblinktitles" request to use domain codes instead of languages
Currently, the "fromsite" property of the payload for "wblinktitles"
request, is constructed by finding replacing the target language with
the source language, inside the "wgWikiID" config variable. However,
this is not working for some edge cases like "nowiki", where the target
language is set to "nb", which is not present inside the "wgWikiID",
resulting to wrong "fromsite" property.

This patch fixes this issue by using the wiki domain codes instead of
the source/target languages, to construct the "fromsite" property.

Bug: T362478
Change-Id: I819c3db1bc229de33dabab3ef5611e051bbc991a