AX article footer entrypoint: Fix condition that adds the module
The logic for that condition is fine, but the "in_array" check, that
uses an instance of "Language" class, instead of a simple language
code (string).

This patch fixes this issue, by using the "getCode" method on that
Language class instance, to check if the content language code is
contained inside the enabled languages for this entrypoint.

Bug: T363338
Change-Id: I132ff6a1e7b13fea57c250e50dedd3b99ae3a751