AX article footer entrypoint: Fix to navigate user to confirmation step
Currently, users are wrongfully navigated to the "translation" step of
the AutomaticTranslation special page. This is different than the task
requirements - according to which we should navigate the user to the
confirmation step, and also creates (possibly unnecessary) increase to
MinT and CXserver instances.

This patch fixes this issue, by redirecting to the desired step.

Bug: T363338
Change-Id: I0bbe8196d4a27643f022f8e551e49f39fc458c39