MinT for Readers: Log "view" event after content has been loaded
According to the schema specs:
```
This should be subsequent event to user clicks to view automatic
translation, but should be logged once the translation is loaded.
```

This patch waits for the first content of the lead section to be
translated, then only logging the event. Until now, the event is
logged when the "view translation" step loads.

Bug: T363685
Change-Id: I609cc1a1acb57094936a23c4407b40d8077b55b3