SX: Use new section suggestions endpoint from recommendation API
Until now, the unified dashboard had to send multiple requests, one
per article, to fetch the suggested sections to translated for the
given article from cxserver. However, now the recommendation API
offers a new endpoint that given a language pair and a seed article
it will return several recommendations (articles) at once. This
saves us a lot of requests and also allows us to follow a unified
approach for page and section suggestions handling.

This patch renames the old "fetchSectionSuggestions" api method to
"fetchSectionSuggestion", singular as a single SectionSuggestion
model is returned, and adds a "new" "fetchSectionSuggestions" method
that uses the recommendation API and returns an array of section
suggestion models. Tests are also fixed.

Bug: T370746
Change-Id: I69ff2cefd5579c11bd8b73fcdea3e8e65c1974ba