SX: Use seeds based on user's previous translations/edits
Until now, the suggestion seeds are based on previously published
translations on CX (by any user). However, we need to update the
way suggestion seeds are fetched, so that we get seeds based on
users previous edits. In following patches, other options for
seed fetching will be added ('most popular', 'topics', etc).

To support this feature, this patch:
- removes "suggestionSeedCollections" state variable and the related
getter and mutation. There was no need for it to be a global state
variable, since it's only used for seed retrieval.
- deletes `useSuggestionSeedsInitialize` composable, and moves all
logic related to seeds in a new composable named "useSuggestionSeeds".
Seeds are only part of the fetching of suggestions based on user
previous edits, requests for most popular suggestions, or suggestions
based on a specific topic (or even campaign), don't need a seed. For
this reason, we isolate the seed-related code and only expose a
"getSuggestionSeed" method, that can be used by methods that fetch
user edit-based suggestions.
- Moves the methods that validate page/section suggestions from
"useSuggestionsFetch" composable to a separate composable named
"useSuggestionValidator" for better code organisation.
- Moves the code that specifically fetches edit-based recommendations
to a separate composable, so that "useSuggestionsFetch" composable
only contains higher level code, and this is the place where we can
add new suggestion 'fetchers', to fetch suggestions based on topics,
etc.

Bug: T370746
Change-Id: Ie475a8825232286c44bcf031091b497dfc6c3058