Unified dashboard: Remove published articles from section suggestions list
When the user visits the "Published" dashboard list, we send a request
to fetch available section suggestions for all published articles. We
store these section suggestions inside vuex store -which helps to avoid
extra requests when user re-visits the "Published" dashboard list, however we miss to mark them as non-listable, so that they are not
displayed in the suggestions list.

This patch fixes this issue by setting "isListable" property to false
for all section suggestions for published translations.

Additionally, this patch stops storing dummy section suggestions to the
store, in the case that the "fetch section suggestion" request fails,
as they pollute the store with dummy data and can cause side effects
(e.g. they cause a bug in translation time indicator inside the
confirmation step.). Instead, we cache the titles for which we have
already sent a request, to avoid sending requests for them too, when
users re-visit the "Published" dashboard list.

Bug: T374559
Change-Id: I64db1a16672bac85e411ef699c49118227ff0159