SX: Add useEventLogging composable to replace event logging plugin
Until now, for logging instrumentation events, we used the
"useEventLogging" method from the "plugins/eventlogging/index.js" module.
Although, the naming suggests that this is a composable, in fact that
"useEventLogging" function used the "provide/inject" to return the
"logEvent" method. Although, this works without any issues until now,
it would be preferred to be replaced with a regular composable, as the
current implementation is not consistent with how composables are
defined, and we need to install a plugin to the Vue app, for it to work.
At the same time, "provide/inject" doesn't offer any advantage in this
case.

This patch removes the "plugins/eventlogging/index.js" file and replaces
it with a "composables/useEventLogging.js" module, that offers the same
functionality, in a more consistent "composable" way:
- plugin removed from Vue app
- all occurrences are fixed
- tests are fixed too

Change-Id: I42b30d9ad1f88bc36f609e5b5ff0d241e0575503