SX: Force eslint to include vue files for `npm run lint`
`npm run lint` script is run as a pre-commit hook, to ensure that the
commited code doesn't contain any eslint errors. However, until now,
Vue files were ignored when the script was run, leading to several
linting errors.

This patch modifies the eslint configuration, so that Vue files are
also parsed by eslint, and fixes all existing warnings and errors.

It also adds a new rule inside .eslintignore, to ignore snap files,
as they are automatically created and don't need to be tested for
such issues.

Change-Id: I8f69aa6bc063154b43981982f56943702d8ba9da