SX: Merge "usePageTranslationStart" and "useSectionTranslationStart"
Context:
When SX was introduced the whole workflow relied on a SectionSuggestion
model, which we stored inside vuex store, as "currentSectionSuggestion".
This was the case both the section translations and page translations
(also called lead section translations). For page translations, the
"currentSectionSuggestion" was basically a dummy model, in the sense,
that it only included the source title and the language pair of the
source article. However, moving to the Unified Dashboard, it was
necessary to remove this assumption about a current section suggestion,
and we needed to rely on URL parameters to "store" all we needed for
a translation: the source title, the language pair and the source
section title (if set). Hence, progressively, we removed the current
section suggestion from vuex state and switched to URL params.

However, there are still remainders of the previous approach in our
codebase. One such remainder is the fact that we use two different
composables for page and section suggestions, although these two
cases can be treated uniformly, using the same composable. This is
also obvious, by checking the "usePageTranslationStart" composable,
that internally uses the "useSectionTranslationStart" composable.

To simplify the codebase, this patch removes the redundant composable,
"usePageTranslationStart", and renames "useSectionTranslationStart to
"useTranslationStart", as this single component is enough to support
all our cases.

Change-Id: I9c7e66c37e8cd0afb757ed06d8a71950c3752f3b