SX: Refactor "dashboard_translation_start" instrumentation
Currently, the "dashboard_translation_start" instrumentation event is
only logged when the Confirmation step is mounted. However, according
to the event description inside schema documentation:
Occurs when the user starts a new translation, after picking the
appropriate article and, in section translation, section and confirming
their choice.

https://github.com/wikimedia/schemas-event-secondary/tree/master/jsonschema/analytics/mediawiki/content_translation_event#dashboard_translation_start

That means, we need to log this event:
a. when the user starts a page translation (for mobile devices this will
be a lead section translation, for desktop devices this will be a page
translation in CX),
b. when the user starts a section translation with the section title
already filled inside Confirmation step, i.e. when the "section" URL
parameter is already set.
c. when the user clicks on the "Translate this section" button, inside
"Compare the contents" step, after they have already selected a specific
section to translate.

This patch moves the logging functionality for this event to a separate
composable, and uses it where needed. It also fixes the logging to also
submit the translation source title and the translation source section
title (if present), that were missing.

Finally, this patch uses a local ref variable to hold the value of the
event source for this event, to avoid the need for passing this value
as prop, inside Vue router components.

Change-Id: If8938c85f94d6c1fd348fa2835f1684324f7d7be