SX: Fix "translation_type" for "dashboard_translation_start" event
Currently, the "translation_type" field for all SX events is hardcoded
to "section", inside "logEvent" method. However, this field should
properly set to "section" or "article", depending on whether the current
translation is a section or a page translation.

This patch fixes this issue by setting this field accordingly, for
"dashboard_translation_start" event, to override the default "section"
value.

Bug: T374905
Change-Id: Id2c5d8791fecd259242c18dbe01a65e70fd53164