SX: Fix "translation_type" for "editor_segment_add" event
Currently, the "translation_type" field for all SX events is hardcoded
to "section", inside "logEvent" method. However, this field should
properly set to "section" or "article", depending on whether the current
translation is a section or a page translation.

This patch moves and renames the "useSentenceSelectorInstrument" composable
to "useEditorInstrument" and adds the "logEditorSegmentAddEvent" that
is used to log the "editor_segment_add" event, with all the appropriate
values. Apart from adding other important missing fields, like
"translation_source_title", "translation_id", etc, this patch also fixes
the "translation_type" field, to be filled with the appropriate value,
instead of the default value ("section").

Bug: T374905
Change-Id: I2de67a9054cdc15c8a68e7948bce36ad17548d68