SX dashboard suggestion list: Only display suggestions fetched for listing
Currently, we have the "useSuggestionLoad" composable that is used to load
the section suggestions for a given language pair and source title (if
target article exists), and also load the page metadata for the source
article. Until now, the method returned by this composable accepted a
"shouldStore" argument, which controlled whether we should also include
the loaded suggestion inside the dashboard suggestion lists.

This patch removes this "shouldStore" argument, and never includes the
loaded suggestion inside the dashboard lists. This happens for two reasons:
a. "useSuggestionLoad" is used to load a suggestion, so that we can properly
display its info inside the confirmation step. Messing with the dashboard
lists, doesn't add any particular value -we just save one or two suggestions,
but we still need to send the same amount of requests. It only adds
complexity to the logic.
b. After the latest changes, where we introduced suggestion filters, and
we filter the page suggestions based on them, we get a "suggestionProvider
is undefined" JS type error, when a page suggestion is selected for the
confirmation step. This happens because, when "loadSuggestion" method is
called and "shouldStore" argument is set to true (which is the case for
translations started from dashboard lists), we add a new page suggestion
inside the store, with missing "suggestionProvider" property.

Change-Id: I66af55cb0e8ebd4b3c4a889e97118d12d0c08764