AX MVP: Instrument "open target article" inside view translation
This event is properly instrumented inside "Confirm topic" page of
AutomaticTranslation special page (MinT MVP for readers), but a
similar card containing a link to the target article also exists
inside "View translation" page. However, the instrumentation for
that card is missing.

This patch adds the missing instrumentation for the target article
click, and also adds the proper `action_context` for both cases
where the event should be logged, so that we recognize from which
step the event is being logged.

Bug: T363685
Change-Id: Id310a051599edf983457adc6368b4114109172cf