AX instrumentation: Fix "view automatic_translation" event to be logged
Currently, we watch the "leadSectionTranslation" ref variable, and the
first time it changes, if the variable includes a non-empty array of
strings, we log the "view automatic translation" instrumentation event.

However, this variable is (until now) initialized as an empty string,
and the watcher is triggered before at the start of the translation
initialization, when its value is set to an empty array. Because the
watcher is set to run only once, the instrumentation event is never
actually logged.

This patch fixes this issue, by removing the "once" option from the
watcher, and let it run until the "leadSectionTranslation" variable
contains at least one string. Then, we log the instrumentation event
and manually unregister the watcher.

Bug: T366149
Change-Id: Ie87c8c362bfa02cc2a1d9403db8da0bfcebdad4d