SX: Move SXSuggestionsFilters SFC inside components/CXDashboard directory
Currently, SXSuggestionsFilters.vue file is located inside components
root directory. However, this SFC is only used inside the dashboard, so
it makes more sense for it to live inside components/CXDashboard for
better code organization. Typically, inside the root "components"
directory, we add directories that map to application pages (steps).

This patch moves that file inside CXDashboard directory, and it
slightly changes it to be a dialog component. It also renames it to
indicate that this is a dialog component.

Change-Id: I1eca9ff190ac8f5563bd6218adf6d3edf66744af