Simplify counting of unique section ids for a translation
In the context of counting the number of translated sections
for a translation, we select the translation id and
count discinct of the section id. However, section id
is also part of the select even though it is not
being used and fails the whole query when ONLY_FULL_GROUP_BY
is set to true, which is the default in mediawiki-docker
local environment.

This patch simplifies the query by using a standard
COUNT() function with the fetchField() function.

Change-Id: I89ce036b54d35458b1c34626e69032657ebae3ec