Avoid banner flicker in relevantlanguages entrypoint
Calling the sitematrix api to validate if one of the frequent language
is a valid domain causes the banner to flicker.

One possible solution for this is to load the sitematrix api on page
load, before user clicks on ULS. But this will call the sitematrix API
for every page load where this entrypoint is enabled. It will impact
the overall reading performance of the page. We should not have
unnecessary API calls on page load.

However, mw.uls.getFrequencyLanguages() returns language codes that are
not guaranteed to be valid domains. The browser language or accept language
codes can be invalid wiki domains. So, we need validation for them.

Prepare the the language code array of valid wikipedia domains at server
side and pass it as a json(RL package file) to the client side. Since the config
value is readily available at client side, we can use it to validate the
language codes from mw.uls.getFrequencyLanguages().

SiteMatrix extension will be used to get the valid wikipedia domains.
If that extension is not available, we will return a list of known valid
languages from a hardcoded list. This fallback mechanism is to avoid the
complexities of setting up SiteMatrix extension in local, testing and staging
environments.

Bug: T344028
Change-Id: Ib9de7995c4bac3b6d30a15cf45760d7b38e9a1bd