SXTranslationListLanguageSelector: Allow selected languages to be null
Until now, inside CXSuggestionList SFC we used source/target languages
from application state (vuex store). These languages were passed as
arguments to the SXTranslationListLanguageSelector SFC, which requires
the selected source and target language props to always be strings.
This requirement was always fulfilled until now, as the source and
target languages were initialized with the "en" value inside the state.
Copying from the description of "sourceLanguage" inside application/state
file:
"The initial value is set to "en" (could be set to any other valid
language code), so that it's always a string (and thus it can be properly
used as prop with type "String" in Vue components)."

Now, that we replaced the usage of vuex store as source of truth for the
source and target languages, with the corresponding URL params - that are
initialized as null - the SXTranslationListLanguageSelector throws a
warning, about the props being null. To avoid these warnings, this patch
allows the selectedSourceLanguage/selectedTargetLanguage props of the
SXTranslationListLanguageSelector SFC to be null by default. This approach
was considered by the uploader, as preferable, to initializing the source
and target language ref variables inside "useURLHandler" with random
languages.

Change-Id: I37154f8ebca053118902fcf26c91cfc8a2f3c78e