Do not slice collection suggestions before saving them to vuex store
Currently, the lists of page and section suggestions fetched from the
Recommendation API for "collections" filter, are sliced to get limited to
"numberOfSuggestionsToFetch" suggestions. This approach is followed for
"For you" filter, in order to avoid displaying suggestions based only on
one seed, and it was blindly copied here. However, there is no need to
slice the suggestions for "collections" filter also, as they are not
based on any seed, and it only leads to additional requests, when more
suggestions for this filter are required.

This patch fixes this issue, by stop slicing the suggestions for the
"collection" filter.

Change-Id: Ia5a1e75084fd38341a061b16201ef650177bc708