Display next suggestion slice only when refresh button is clicked
Currently, when user discards a suggestion and another suggestion
is already available to be displayed instead, the suggestions are
refreshed, and the next suggestions slice is displayed. The same
happens if the user selects a previously selected filter, for which
the first suggestions slice is already complete (i.e. 3 suggestions
already exist).

This patch fixes these issues, by refreshing the suggestions (i.e.
displaying the next suggestions slice), only when the user clicks
on the "Refresh suggestions" button. At the same time, the patch
fixes the loading indicator inside suggestions list, to be displayed
only if the current slice is not full, as otherwise, there is no need
to indicate to the user that more suggestions are being fetched.

Bug: T381146
Bug: T381147
Change-Id: I72da3e12dd3e20ca7ed96bce9db5a9357bdec23e