Add state for "active-list" URL parameter
Until now, the "active-list" URL parameter, that holds the value for the
active dashboard tab ("suggestions", "in progress", "published"), is
treated differently from other SX URL params, and there is no state
variable where it's value is stored.

This patch adds a state variable and a setter to treat this URL parameter
the same way as we treat all the other standard SX URL params.

Some refactoring in useURLHandler composable also done.

Change-Id: I31f7401a84ca0f9e0dfc7f8a01155275384788da