i18n: delete messages for article topics being moved to WikimediaMessages
All the topic messages are being moved to the WikimediaMessages
extension for easier reuse. The first step of the migration plan, as
described in T380825#10481384, is to move all message definitions
verbatim, i.e. remove all messages from here and add them to
WikimediaMessages (see patches with the same topic, which should be
merged together with this one).

Most messages are being taken from the GrowthExperiments extension,
except for `-history-and-society` and `-science-technology-and-math`
that are instead taken from here.

Bug: T380825
Depends-On: Ic741bf5b5e52365c7a7f09330b6ae694b61b028a
Change-Id: I50debe840c45acc7f1a3ca7d8d3da7dba8ae2416