Adjust language selector to the corresponding screen size
Currently, the language selector is positioned at the same position for
both mobile and desktop devices. However, according to the design
specifications, in desktop devices the language selector should be
placed next to the dashboard tab title. This patch adjusts the layout
for desktop devices, so that the language selector is placed
accordingly.

Some other minor changes:
- Removed unused `useDevice` from CXTranslationWorkPublished SFC
- Replaced usage of MwButton with CdxButton, and added the "expand" icon
inside SXTranslationListLanguageSelector SFC
- Some other minor visual adjustments so that the layout is closer to
the design.

Bug: T351734
Change-Id: I0f65431305c054a7977ccd3444346b286d0c6bfc