Unified Dashboard: Do not show the cross-device warning on desktop
Confirmation button inside the dialog is also fixed to take all the
available space.

Note for testing: Checkout on master, and start a translation for
an article on desktop from the unified dashboard. Make some changes
and wait for them to be saved. Then, go back to the dashboard and
click on "New translation" button. Search for the same article,
select it for translation and click on the "Continue translation"
button. You should get the "Continue translation on mobile" dialog.

Now checkout to this patch, and search for the same article again.
Clicking on the "Continue translation" button should not show you
the dialog.

Bug: T385795
Change-Id: I6bc4d9623c6610921318f791d363bf8361c51270