SX search step: Check if suggested languages are supported
Currently, we consider as valid, any language that has a valid autonym.
However, this is not enough, as not all languages with valid autonyms
are supported for Content Translation.

This patch fixes this issue by actually checking if each suggested
languages belongs to the list of supported language codes, before
considering it as valid.

Change-Id: Ide49c33ab9361879b03619003d2b85eeaad22620