SX: Clear all translation URL params when going back to dashboard
Currently, we have 3 steps that offer the possibility to the user
to return to the dashboard: "Confirm a translation", "Pick a section"
and "Pick a sentence" steps. All of them face issues regarding the
way they "clear" the URL params. More specifically:
- in confirmation step, we preserve URL params like source/target
languages and filters, but we only clear the "page" URL parameter,
leaving "section" and "draft" URL parameters intact.
- in section selection and sentence selection steps, we clear all
URL parameters, including those that are supposed to be preserved
like source/target languages and filters.

This patch renames the "clearURLParameters" method that is used
to clear the necessary URL parameters in these - and only these -
places, and fixes it to only clear the appropriate params (page,
section, draft), instead of all of them.

Bug: T385287
Change-Id: I4cdac785da009fc0310ea2f0aea4c90d146adefb