Search for an article: Use previous edits in source wiki for suggestions
Until now, the "Search for an article" step displayed suggestions in the
"Recently edited by you" section based on previously published
translations in the current language pair, instead of the most recently
edited articles in the source language.

This patch fixes the issue by ensuring that the suggestions in the
"Recently edited by you" section are truly based on the most recently
edited articles in the source language.

Bug: T282430
Change-Id: Ic92dcc82956f38215437c28b19e22e842d88e971