Compare contents header: Fix spacing to match specifications
Until now, in order to position the elements inside the header of the
"Compare contents" step, we use <mw-row> and <mw-col> elements with
specified column-width (as percentage). This is quite restrictive as
it fixes the width taken by each element inside the row, and it doesn't
allow the elements to "adjust" and take as much width as needed.

Because of this, the spacing between the different elements inside the
header doesn't match the design specifications. This patch fixes this
issue by using a simple "flexbox" approach.

Bug: T386696
Change-Id: Idb4a4024068719d89635c108e8811df830c28cdc