Unified Dashboard: Prevent unnecessary loading indicator on return
Currently, when the user navigates to the confirmation step (or any
other step), and then back to the dashboard, a loading indicator is
being displayed and the "Refresh suggestions" button is missing.

This patch fixes this issue by removing the redundant mutation for
`isSectionSuggestionsFetchPending` state variable.

Screencast before fix: https://phabricator.wikimedia.org/F58423260

Fixes: I779bccccd70070adfca271fdc05b4a9cc0db76dd
Change-Id: I40d9a623df7a57f5e50f8480cf208c8993f124ef