SX: Introduce `useTranslationConfirm` composable
The code handling translation confirmation —redirecting the user to the
desktop/mobile editor or the quick tutorial step if needed— is almost
identical in both the confirmation step and the "Compare the contents"
step.

To avoid code duplication, increase maintainability and for better code
organisation this patch introduces the `useTranslationConfirm`
composable that will be used in all 3 cases that we currently need this
functionality.

Also removed unused `isDesktop` variable from SXSectionSelector.

Bug: T386344
Change-Id: I7d5b629a6c28c7b0ec1074e088938ac0324f6e14