Remove comma after "Please"
All the messages in MediaWiki core and in all the extensions
used by Wikimedia don't use a comma after the word "Please"
that begins a sentence. This comma is generally unnecessary,
and English style guides recommend using it only when more
emphasis is needed, and it is not neeeded here.

Change-Id: Ie65e152e0fff0744c9e74d01b587100a8b737fdc