Respect selected tab from desktop dashboard URLs
The desktop dashboard store selected tab in the hash
part of the URL (e.g. #suggestions) but the unified
dashboard stores it in 'active-list' query parameter.

Initialize the selected tab with the hash when the
'active-list' parameter is not specified in order
to be compatible with existing CX dashboard URLs
that are used throughout projects.

Bug: T386896
Change-Id: I3d9f7092a4f7cff01702d2fd4ed20b6cca4e7f56